requirements.txt
torch==2.0.0
torchvision==0.15.1
open_clip_torch==2.18.0
addict==2.4.0
yapf==0.33.0
pandas
scipy
transformers==4.28.1
sentence-transformers==2.2.2